third_party/WebKit/LayoutTests/external/wpt/fullscreen/interfaces-expected.txt

This is a testharness.js-based test.
PASS Test driver 
FAIL Document interface: attribute fullscreenEnabled assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
FAIL Document interface: attribute fullscreen assert_true: The prototype object must have a property "fullscreen" expected true got false
FAIL Document interface: operation exitFullscreen() assert_unreached: Throws "TypeError: Illegal invocation" instead of rejecting promise Reached unreachable code
PASS Document interface: attribute onfullscreenchange 
PASS Document interface: attribute onfullscreenerror 
FAIL Document interface: attribute fullscreenElement assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
FAIL Document must be primary interface of new Document assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Stringification of new Document assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Document interface: new Document must inherit property "fullscreenEnabled" with the proper type (29) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Document interface: new Document must inherit property "fullscreen" with the proper type (30) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Document interface: new Document must inherit property "exitFullscreen" with the proper type (31) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Document interface: new Document must inherit property "onfullscreenchange" with the proper type (32) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Document interface: new Document must inherit property "onfullscreenerror" with the proper type (33) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Document interface: new Document must inherit property "fullscreenElement" with the proper type (35) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL ShadowRoot interface: attribute fullscreenElement assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
FAIL Element interface: operation requestFullscreen() assert_unreached: Throws "TypeError: Illegal invocation" instead of rejecting promise Reached unreachable code
PASS Element must be primary interface of document.createElementNS(null, "test") 
PASS Stringification of document.createElementNS(null, "test") 
PASS Element interface: document.createElementNS(null, "test") must inherit property "requestFullscreen" with the proper type (34) 
Harness: the test ran to completion.
